% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcolors.R
\name{display_all_jcolors}
\alias{display_all_jcolors}
\title{Display all jcolors}
\usage{
display_all_jcolors()
}
\arguments{
\item{palette}{Character string indicating a palette of colors.}
}
\description{
Creates different vectors of related colors that may be useful for figures.
}

