% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contin_jcolors.R
\name{display_all_jcolors_contin}
\alias{display_all_jcolors_contin}
\title{Display every jcolors_contin palette}
\usage{
display_all_jcolors_contin()
}
\description{
displays all of the continuous jcolors palettes
}
\examples{
display_all_jcolors_contin()
}
