% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcolors.R
\name{display_jcolors}
\alias{display_jcolors}
\title{Display jcolors}
\usage{
display_jcolors(palette = c("default"))
}
\arguments{
\item{palette}{Character string indicating a palette of colors.}
}
\value{
Vector of character strings representing the chosen palette of colors.
}
\description{
Creates different vectors of related colors that may be useful for figures.
}

