% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contin_jcolors.R
\name{display_jcolors_contin}
\alias{display_jcolors_contin}
\title{Display jcolors_contin}
\usage{
display_jcolors_contin(
  palette = c("default", "pal2", "pal3", "pal4", "pal10", "pal11", "pal12", "rainbow")
)
}
\arguments{
\item{palette}{Character string indicating a palette of colors.}
}
\description{
displays the continuous jcolors palettes
}
\examples{
display_jcolors_contin()
}
