% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcolors.R
\name{jcolors}
\alias{jcolors}
\title{Vectors of colors for figures}
\usage{
jcolors(set = c("default"))
}
\arguments{
\item{set}{Character string indicating a set of colors.}
}
\value{
Vector of character strings representing the chosen set of colors.
}
\description{
Creates different vectors of related colors that may be useful for figures.
}
\examples{
par(mar=c(0.6,5.1,0.6,0.6))
plot(0, 0, type="n", xlab="", ylab="", xlim=c(0, 3), ylim=c(2.5, 0), yaxs="i",
     xaxt="n", yaxt="n", xaxs="i")
axis(side=2, at=1:2, c("default", "highlight"), las=1)

def <- jcolors("default")
points(seq(along = def), rep(1, length(def)), pch = 22, bg = def, cex = 5)
}

