% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcolors.R
\name{scale_color_jcolors}
\alias{scale_color_jcolors}
\alias{scale_colour_jcolors}
\alias{scale_fill_jcolors}
\title{jcolors color scales}
\usage{
scale_color_jcolors(palette = c("default", "pal2", "pal3"), ...)

scale_colour_jcolors(palette = c("default", "pal2", "pal3"), ...)

scale_fill_jcolors(palette = c("default", "pal2", "pal3"), ...)
}
\arguments{
\item{palette}{Character string indicating a palette of colors.}

\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
jcolors color scales
}
\examples{
library(ggplot2)
data(morley)

pltl <- ggplot(data = morley, aes(x = Run, y = Speed,
group = factor(Expt),
colour = factor(Expt))) +
    geom_line(size = 2) +
    theme_bw() +
    theme(panel.background = element_rect(fill = "grey97"),
          panel.border = element_blank())

pltd <- ggplot(data = morley, aes(x = Run, y = Speed,
group = factor(Expt),
colour = factor(Expt))) +
    geom_line(size = 2) +
    theme_bw() +
    theme(panel.background = element_rect(fill = "grey15"),
          panel.border = element_blank(),
          panel.grid.major = element_line(color = "grey45"),
          panel.grid.minor = element_line(color = "grey25"))

pltl + scale_color_jcolors(palette = "default")

pltd + scale_color_jcolors(palette = "default")


}

